+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.2.1] - 2023-08-10
 
 - Added support for schema updates to `market` package models to include DMA and CBSA details
+- Adjusted the `options.AddFilter` method to be variadic to support passing multiple values easily
 
 ## [0.2.0] - 2023-08-10